Moving ParseEventMarker to akka.stream.alpakka.xml package
